用户登陆注册接口

+++ src/main/java/im/peng/ceramic/ceramic_sell/model/po/UserInfo.java
@@ -5,20 +5,21 @@ import lombok.Data;
 
 @Data
 public class UserInfo implements Serializable {
+
     private String id;
 
-    private Integer gmtCreate;
+    private Long gmtCreate;
 
     private Long gmtModify;
 
     /**
-    * 小程序中微信用户身份唯一标识
-    */
+     * 小程序中微信用户身份唯一标识
+     */
     private String vxOpenid;
 
     /**
-    * 会员卡id
-    */
+     * 会员卡id
+     */
     private String memberId;
 
     private static final long serialVersionUID = 1L;
